+++ packages/studio-base/src/App.tsx
@@ -20,7 +20,7 @@ import AnalyticsProvider from "./context/AnalyticsProvider";
 import AppConfigurationContext, { IAppConfiguration } from "./context/AppConfigurationContext";
 import { AssetsProvider } from "./context/AssetsContext";
 import ConsoleApiContext from "./context/ConsoleApiContext";
-import ExtensionLoaderContext, { ExtensionLoader } from "./context/ExtensionLoaderContext";
+import ExtensionLoaderContext from "./context/ExtensionLoaderContext";
 import { HoverValueProvider } from "./context/HoverValueContext";
 import LayoutStorageContext from "./context/LayoutStorageContext";
 import ModalHost from "./context/ModalHost";
@@ -40,6 +40,7 @@ import PanelCatalogProvider from "./providers/PanelCatalogProvider";
 import UserProfileLocalStorageProvider from "./providers/UserProfileLocalStorageProvider";
 import { LaunchPreference } from "./screens/LaunchPreference";
 import ConsoleApi from "./services/ConsoleApi";
+import { ExtensionLoader } from "./services/ExtensionLoader";
 import { ILayoutStorage } from "./services/ILayoutStorage";
 import URDFAssetLoader from "./services/URDFAssetLoader";
