+++ packages/studio-base/src/services/IExtensionStorage.ts
@@ -2,7 +2,7 @@
 // License, v2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
-import { ExtensionInfo } from "@foxglove/studio-base/context/ExtensionLoaderContext";
+import { ExtensionInfo } from "@foxglove/studio-base/services/ExtensionLoader";
 
 export type StoredExtension = {
   info: ExtensionInfo;
