+++ packages/studio-base/src/services/IdbExtensionLoader.ts
@@ -5,8 +5,12 @@
 import JSZip from "jszip";
 
 import Log from "@foxglove/log";
-import { ExtensionInfo, ExtensionLoader, IExtensionStorage } from "@foxglove/studio-base";
-import { ExtensionNamespace } from "@foxglove/studio-base/context/ExtensionLoaderContext";
+import {
+  ExtensionInfo,
+  ExtensionLoader,
+  ExtensionNamespace,
+} from "@foxglove/studio-base/services/ExtensionLoader";
+import { IExtensionStorage } from "@foxglove/studio-base/services/IExtensionStorage";
 
 import { IdbExtensionStorage } from "./IdbExtensionStorage";
