+++ packages/studio-base/src/services/IdbExtensionStorage.ts
@@ -5,7 +5,11 @@
 import * as IDB from "idb/with-async-ittr";
 
 import Log from "@foxglove/log";
-import { StoredExtension, IExtensionStorage, ExtensionInfo } from "@foxglove/studio-base";
+import { ExtensionInfo } from "@foxglove/studio-base/services/ExtensionLoader";
+import {
+  IExtensionStorage,
+  StoredExtension,
+} from "@foxglove/studio-base/services/IExtensionStorage";
 
 const log = Log.getLogger(__filename);
